# api/docs/swagger.yaml
openapi: 3.0
info:
  title: Bank API
  description: API for managing accounts and transactions
  version: 1.0.0
servers:
  - url: 'http://localhost:3000'
    description: Local development server

paths:
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: string
                password:
                  type: string
              required:
                - user
                - password
      responses:
        '201':
          description: User registered successfully

  /auth/login:
    post:
      summary: Login a user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: string
                password:
                  type: string
              required:
                - user
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string

  /accounts:
    post:
      summary: Create a new account
      tags:
        - Accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: string
                currency:
                  type: string
                description:
                  type: string
                balance:
                  type: number
              required:
                - user
                - currency
                - balance
      responses:
        '201':
          description: Account created successfully

  /accounts/{user}:
    get:
      summary: Get account details
      tags:
        - Accounts
      parameters:
        - name: user
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account details
        '404':
          description: User not found